+++ functions/Inputs.php
@@ -101,7 +101,12 @@ function TextInput( $value, $name, $title = '', $extra = '', $div = true )
 	if ( mb_strpos( $extra, 'size=' ) === false )
 	{
 		// Max size is 32 (more or less 300px)
-		$extra .= $value != '' ? ' size="' . min( mb_strlen( $value ), 32 ) . '"' : ' size="12"';
+		$size = min( mb_strlen( (string) $value ), 32 );
+
+		// Min size is 2 (more or less 35px)
+		$size = max( $size, 2 );
+
+		$extra .= $value != '' ? ' size="' . $size . '"' : ' size="12"';
 	}
 
 	// Specify input type via $extra (email,...).